backend/requirements.txt
Django
djangorestframework
django-cors-headers
djangorestframework-simplejwt
python-decouple
Pillow
channels
channels-redis
daphne
redis
requests
celery
django-celery-beat
scikit-learn
pandas
numpy
redis
cloudinary
twilio
django-environ
dnspython
email-validator
pyjwt
bcrypt
python-dotenv
drf-spectacular 
drf-spectacular-sidecar
psycopg2-binary

